In maps: Distance value type changed from Integer to Double Reason: This interface can now also be applied to real map data such as Open Street Map data, where node positions are sometimes really near and sometimes very far away from each other.

+++ src/aima/gui/applications/search/map/AbstractMapAgentController.java
@@ -68,7 +68,7 @@ public abstract class AbstractMapAgentController extends AgentAppController {
 			Agent agent = (Agent) agents.get(0);
 			String status = (String) agent
 					.getAttribute(DynAttributeNames.AGENT_STATUS);
-			Integer travelDistance = (Integer) agent
+			Double travelDistance = (Double) agent
 					.getAttribute(DynAttributeNames.AGENT_TRAVEL_DISTANCE);
 			StringBuffer statusMsg = new StringBuffer("Task ");
 			if (status != null)
